+ createSimplifiedTransaction
